🐛  Fix list dead link by moving to own folder
